requirements.txt
neo4j
langgraph
langchain-core
langchain-community
langchain-openai
langchain-neo4j
openai
python-dotenv
requests
numpy
pandas
pydantic
typing_extensions
arize-phoenix
arize-phoenix-otel
openinference-instrumentation-langchain
openinference-instrumentation-openai
opentelemetry-sdk
opentelemetry-exporter-otlp
